% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsetreg-package.R
\docType{package}
\name{offsetreg-package}
\alias{offsetreg}
\alias{offsetreg-package}
\title{offsetreg: Tidymodels engines with support for offset terms}
\description{
This package provides engines compatible with the tidymodels ecosystem that support offset terms.
}
\author{
\strong{Maintainer}: Matt Heaphy \email{mattrmattrs@gmail.com}

}
\keyword{internal}
