% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsetreg-package.R
\docType{package}
\name{offsetreg-package}
\alias{offsetreg}
\alias{offsetreg-package}
\title{offsetreg: Tidymodels engines with support for offset terms}
\description{
This package provides functions that extend the tidymodels ecosystem \url{https://www.tidymodels.org/} to enable usage of predictive models with offset terms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattheaphy/offsetreg/}
  \item \url{https://mattheaphy.github.io/offsetreg/}
}

}
\author{
\strong{Maintainer}: Matt Heaphy \email{mattrmattrs@gmail.com}

}
\keyword{internal}
